Refactor checklist instantation / diffing
This removes the Todoist/TripIt dependencies from the diff logic and adds tests.
The Todoist logic largely moves into the todoist package, and the TripIt logic
(minor) is moved to main.

Checklist is now expanded into Tasks, which are members of a tasks.Project.
Project instances can be diffed and the diffs are represented by a list of
Diffs. These are converted to/from Todoist internal types.

This also adds:
  * v7 API support for Todoist (was a no-op)
  * Better error detection for Todoist write commands