Refine trip start and end times using TripIt's AirObject
TripIt Trip objects have resolution to the day, and tasks were created using
this resolution. This did not permit checklist items within a few hours of
a trip (e.g; checkin to your flight). This change sets ActualStartDate and
ActualEndDate on a Trip based on the AirObjects in the TripIt response.

As a result, this code started needing to handle timezones. The implicit
use of UTC everywhere has been changed for either trip timezone time
(if provided from the AirObject) or local time (for task cutoffs, task
generation). UTC is still used within Task objects and the fields are
renamed accordingly.