Fix a bug with Tripit
If you only have one Trip upcoming, then Tripit returns the Trip object inline
and not a list-of-Trips as it does when you have >1 trip.

This introduces a TripitSingleTripResponse thats used as a fall-through object
to unmarshal into. This was chosen for clarity over attempting to "fix" the JSON
inline to make it compatible with the pre-existing TripitResponse message.