Tidy up the repo to ~modern standards (#4)
* Tidy up the repo to ~modern standards

While here, upgrade go to 1.20, tidy the modules, and upgrade the direct
and indirect dependencies. Also update the GitHub workflow.

Most of this is just reorganising code to match current custom. This has the
benefit that it makes having 2 binaries a bit easier to manage (and
theoretically will make it easier to separate out the API verification
tool from tripist itself, later on).